# swagger/taxii-client-api-0.5.yml
openapi: 3.0.1
info:
  title: TAXII Client API
  description: TAXII Client API (2.1)
  contact:
    name: StephenOTT
    url: http://github.com/stephenott/TAXII-Server
    email: http://github.com/stephenott/TAXII-Server
  license:
    name: --
    url: http://github.com/stephenott
  version: "0.5"
paths:
...
  /taxii2:
    get:
      summary: Get information about the TAXII Server and any advertised API Roots
      description: This Endpoint provides general information about a TAXII Server,
        including the advertised API Roots. It's a common entry point for TAXII Clients
        into the data and services provided by a TAXII Server.
      operationId: discovery
      parameters:
      - name: Accept
        in: header
        required: true
        schema:
          type: string
      responses:
        200:
          description: The request was successful
          headers:
            Content-Type:
              description: Always value of application/taxii+json;version=2.1
              required: true
          content:
            application/taxii+json;version=2.1:
              schema:
                $ref: '#/components/schemas/discovery'
        400:
          description: The server did not understand the request or filter parameters
          content:
            application/taxii+json;version=2.1:
              schema:
                $ref: '#/components/schemas/error'
        401:
          description: The client needs to authenticate
          content:
            application/taxii+json;version=2.1:
              schema:
                $ref: '#/components/schemas/error'
        403:
          description: The client does not have access to this resource
          content:
            application/taxii+json;version=2.1:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: The Discovery service is not found, or the client does not
            have access to the resource
          content:
            application/taxii+json;version=2.1:
              schema:
                $ref: '#/components/schemas/error'
        406:
          description: The media type provided in the Accept header is invalid
          content:
            application/taxii+json;version=2.1:
              schema:
                $ref: '#/components/schemas/error'
      security:
      - basicAuth: []
components:
  schemas:
...
    Dictionary:
      type: object
      properties:
        dictionary:
          type: object
          additionalProperties: true
      description: TAXII Dictionary
    error:
      type: object
      properties:
        title:
          type: string
          description: A human readable plain text title for this error.
        description:
          type: string
          description: A human readable ​plain text​ description that gives details
            about the error or problem that was encountered by the application.
          nullable: true
        error_id:
          type: string
          description: An identifier for this particular error instance. A TAXII Server
            might choose to assign each error occurrence its own identifier in order
            to facilitate debugging.
          nullable: true
        error_code:
          type: string
          description: The error code for this error type. A TAXII Server might choose
            to assign a common error code to all errors of the same type. Error codes
            are application-specific and not intended to be meaningful across different
            TAXII Servers​.
          nullable: true
        http_status:
          type: string
          description: The HTTP status code applicable to this error. If this property
            is provided it ​MUST​ match the HTTP status code found in the HTTP header.
          nullable: true
        external_details:
          type: string
          description: A URL that points to additional details. For example, this
            could be a URL pointing to a knowledge base article describing the error
            code. Absence of this property indicates that there are no additional
            details.
          nullable: true
        details:
          $ref: '#/components/schemas/Dictionary'
      description: The ​error​ message is provided by TAXII Servers in the response
        body when returning an HTTP error status and contains more information describing
        the error, including a human-readable ​title​ and description​, an ​error_code​
        and ​error_id​, and a ​details​ structure to capture further structured information
        about the error. All of the properties are application-specific, and clients
        shouldn't assume consistent meaning across TAXII Servers even if the codes,
        IDs, or titles are the same.
...
    discovery:
      type: object
      properties:
        title:
          type: string
          description: A human readable plain text name used to identify this server.
        description:
          type: string
          description: A human readable plain text description for this server.
          nullable: true
        contact:
          type: string
          description: The human readable plain text contact information for this
            server and/or the administrator of this server.
          nullable: true
        default:
          type: string
          nullable: true
        api_roots:
          type: array
          description: A list of URLs that identify known API Roots. This list MAY​
            be filtered on a per-client basis.  API Root URLs ​MUST​ be HTTPS absolute
            URLs or relative URLs. API Root relative URLs ​MUST​ begin with a single
            "/" character and ​MUST​ N​ OT​ begin with "//" or "../". API Root URLs
            ​MUST NOT​ contain a URL query component.
          nullable: true
          items:
            type: string
      description: The ​discovery​ resource contains information about a TAXII Server,
        such as a human-readable ​title​, description​, and ​contact​ information,
        as well as a list of API Roots that it is advertising. It also has an indication
        of which API Root it considers the ​default​, or the one to use in the absence
        of another information/user choice.
  securitySchemes:
    basicAuth:
      type: http
      description: Basic Authentication
      scheme: Authorization